feat: implementa context api entre páginas home e history
