+++ src/pages/History/index.tsx
@@ -1,6 +1,10 @@
+// import { useContext } from 'react'
 import { HistoryContainer, HistoryList, Status } from './styles'
+// import { CyclesContext } from '../../contexts/CyclesContext'
 
 export function History() {
+  // const { cycles } = useContext(CyclesContext)
+
   return (
     <HistoryContainer>
       <h2>Meu histórico</h2>
