+++ src/pages/Home/Countdown/index.tsx
@@ -1,7 +1,7 @@
 import { useContext, useEffect } from 'react'
 import { differenceInSeconds } from 'date-fns'
 import { CountdownContainer, Separator } from './styles'
-import { CyclesContext } from '..'
+import { CyclesContext } from '../../../contexts/CyclesContext'
 
 export function Countdown() {
   const {
