+++ src/pages/Home/NewCycleForm/index.tsx
@@ -1,8 +1,8 @@
 import { useContext } from 'react'
+import { useFormContext } from 'react-hook-form'
 
 import { FormContainer, MinutesAmountInput, TaskInput } from './styles'
-import { CyclesContext } from '..'
-import { useFormContext } from 'react-hook-form'
+import { CyclesContext } from '../../../contexts/CyclesContext'
 
 export function NewCycleForm() {
   const { activeCycle } = useContext(CyclesContext)
